// src/types/types.tsx
export type Winner = {
    name: string;
    generation: string;
    winBattlesCount: number;
  };
  
  export type Judge = {
    name: string;
    generation: string;
    judgesCount: number;
  };
  
  export type Performer = {
    name: string;
    generation: string;
    performancesCount: number;
  };
  